Make BuildScanPluginPerformanceTest baseline-aware
