+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/fixture/CrossVersionPerformanceTestRunner.groovy
@@ -17,9 +17,6 @@
 package org.gradle.performance.fixture
 
 import com.google.common.annotations.VisibleForTesting
-import com.google.common.base.Splitter
-import groovy.transform.CompileStatic
-import groovy.transform.TypeCheckingMode
 import org.apache.commons.io.FileUtils
 import org.gradle.integtests.fixtures.RepoScriptBlockUtil
 import org.gradle.integtests.fixtures.executer.GradleDistribution
@@ -47,17 +44,17 @@ import org.junit.Assume
 
 import java.util.function.Consumer
 import java.util.function.Function
-import java.util.regex.Pattern
 
 import static org.gradle.integtests.fixtures.RepoScriptBlockUtil.gradlePluginRepositoryMirrorUrl
+import static VersionResolver.isRcVersionOrSnapshot
+import static VersionResolver.resolveBaselineVersions
+import static VersionResolver.resolveVersion
 import static org.gradle.test.fixtures.server.http.MavenHttpPluginRepository.PLUGIN_PORTAL_OVERRIDE_URL_PROPERTY
 
 /**
  * Runs cross version performance tests using Gradle profiler.
  */
 class CrossVersionPerformanceTestRunner extends PerformanceTestSpec {
-
-    private static final Pattern COMMA_OR_SEMICOLON = Pattern.compile('[;,]')
 
     private final IntegrationTestBuildContext buildContext
     private final DataReporter<CrossVersionPerformanceResults> reporter
@@ -175,42 +172,10 @@ class CrossVersionPerformanceTestRunner extends PerformanceTestSpec {
         perVersion
     }
 
-    private static String resolveVersion(String version, ReleasedVersionDistributions releases) {
-        switch (version) {
-            case 'last':
-                return releases.mostRecentRelease.version.version
-            case 'nightly':
-                return LatestNightlyBuildDeterminer.latestNightlyVersion
-            case 'defaults':
-                throw new IllegalArgumentException("'defaults' shouldn't be used in target versions.")
-            default:
-                def releasedVersion = findRelease(releases, version)
-                if (releasedVersion) {
-                    return releasedVersion.version.version
-                } else if (isRcVersionOrSnapshot(version)) {
-                    // for snapshots, we don't have a cheap way to check if it really exists, so we'll just
-                    // blindly add it to the list and trust the test author
-                    // Only active rc versions are listed in all-released-versions.properties that ReleasedVersionDistributions uses
-                    return version
-                } else {
-                    throw new RuntimeException("Cannot find Gradle release that matches version '$version'")
-                }
-        }
-    }
-
     @VisibleForTesting
     static Iterable<String> toBaselineVersions(ReleasedVersionDistributions releases, List<String> targetVersions, String minimumBaseVersion) {
-        List<String> versions
         def overrideBaselinesProperty = System.getProperty('org.gradle.performance.baselines')
-        if (overrideBaselinesProperty) {
-            versions = resolveOverriddenVersions(overrideBaselinesProperty, targetVersions)
-        } else {
-            versions = targetVersions
-        }
-
-        if (versions.contains("none")) {
-            return []
-        }
+        def versions = resolveBaselineVersions(overrideBaselinesProperty, targetVersions)
 
         LinkedHashSet<String> resolvedVersions = versions.collect { resolveVersion(it, releases) } as LinkedHashSet<String>
 
@@ -239,34 +204,6 @@ class CrossVersionPerformanceTestRunner extends PerformanceTestSpec {
 
     private static boolean versionMeetsLowerBaseVersionRequirement(String targetVersion, String minimumBaseVersion) {
         return minimumBaseVersion == null || GradleVersion.version(targetVersion).baseVersion >= GradleVersion.version(minimumBaseVersion)
-    }
-
-    @CompileStatic(TypeCheckingMode.SKIP)
-    private static boolean isRcVersionOrSnapshot(String version) {
-        GradleVersion versionObject = GradleVersion.version(version)
-        // there is no public API for checking for RC version, this is an internal way
-        return versionObject.snapshot || versionObject.stage?.stage == 3
-    }
-
-    private static List<String> resolveOverriddenVersions(String overrideBaselinesProperty, List<String> targetVersions) {
-        def versions = Splitter.on(COMMA_OR_SEMICOLON)
-            .omitEmptyStrings()
-            .splitToList(overrideBaselinesProperty)
-        versions.collectMany([] as Set) { version -> version == 'defaults' ? targetVersions : [version] } as List
-    }
-
-    private static GradleDistribution findRelease(ReleasedVersionDistributions releases, String requested) {
-        GradleDistribution best = null
-        for (GradleDistribution release : releases.all) {
-            if (release.version.version == requested) {
-                return release
-            }
-            if (!release.version.snapshot && release.version.baseVersion.version == requested && (best == null || best.version < release.version)) {
-                best = release
-            }
-        }
-
-        best
     }
 
     private void runVersion(String displayName, GradleDistribution dist, File workingDir, MeasuredOperationList results) {
